+++ ollama_usage/__init__.py
@@ -1,4 +1,19 @@
 from ollama_usage.scraper import get_usage
-from ollama_usage.exceptions import OllamaUsageError, AuthError, ParseError, NetworkError
+from ollama_usage.exceptions import (
+    OllamaUsageError,
+    AuthError,
+    ParseError,
+    NetworkError,
+    BrowserNotFoundError,
+    UnsupportedOSError,
+)
 
-__all__ = ["get_usage", "OllamaUsageError", "AuthError", "ParseError", "NetworkError"]
+__all__ = [
+    "get_usage",
+    "OllamaUsageError",
+    "AuthError",
+    "ParseError",
+    "NetworkError",
+    "BrowserNotFoundError",
+    "UnsupportedOSError",
+]